One approach at addressing https://github.com/JuliaLang/julia/issues/33490.

As described here: https://github.com/JuliaLang/julia/issues/9224#issuecomment-889549789, this PR has Profile.jl randomly permute the order in which it samples the threads for profiling. It still samples all threads on every pause, but it maintains an array with the order for which to sample the threads, and shuffles that array before each sample, then samples the threads according to that order.

The hope is that this will reduce the likelihood of introducing artificial contention into the program, by avoiding pausing the threads in a consistent order, which can cause a pileup of threads if they share a mutex, as described here:
https://github.com/JuliaLang/julia/issues/33490#issue-503572179


--------------------------

We seem to think that the best approach is to change the profiler to:
 - use a random sample interval
 - only sample one thread each time, chosen randomly
But this PR is a good step in the right direction.


One potential concern is that it will unacceptably reduce performance of profiling, so that remains to be seen.

I did some small measurements, and shuffling the array seems _reasonably_ fast:
```julia
julia> const x = [1,2,3]
3-element Vector{Int64}:
 1
 2
 3

julia> const seed = Ref(0)
Base.RefValue{Int64}(0)

julia> @btime @ccall jl_shuffle_int_array_inplace(pointer(x)::Ptr{Int}, length(x)::Csize_t, seed::Ptr{Int})::Cvoid
  17.172 ns (0 allocations: 0 bytes)
```
```julia
julia> const x = collect(1:100)
100-element Vector{Int64}:
   1
   ⋮
 100

julia> const seed = Ref(0)
Base.RefValue{Int64}(0)

julia> @btime @ccall jl_shuffle_int_array_inplace(pointer(x)::Ptr{Int}, length(x)::Csize_t, seed::Ptr{Int})::Cvoid
  710.507 ns (0 allocations: 0 bytes)
```