@NHDaly 2021-07-30
@vtjnash: is this more or less what you had in mind? Thanks!