@vtjnash 2021-07-30
Note that I don't think this will not reduce the appearance of artificial contention, since they are still likely to pile up against mutexes held while stopped for the profile, just that the contention will now be evenly distributed over threads, instead of being proportional to thread id.