@NHDaly 2021-07-30
I don't know if I have the power to invoke a benchmark suite run.. Can you do that? Do i just run NanoSoldier? Are there profiling-specific benchmarks to run, here?