@NHDaly 2021-07-30
Actually, it looks like there aren't any benchmarks for `Profile`? 🤔
https://github.com/JuliaCI/BaseBenchmarks.jl/search?q=profile