@NHDaly 2021-08-01
Okay excellent, it looks like there's no measurable perf impact here:

Before (Julia 1.7):
```julia
julia> Threads.nthreads()
100

julia> function myfunc()
           A = rand(200, 200, 400)
           maximum(A)
       end
myfunc (generic function with 1 method)

julia> @btime (Profile.init(n=Int(1e8), delay=0.0001); Profile.clear(); @profile myfunc())
  40.615 ms (2 allocations: 122.07 MiB)
0.9999999702097592

julia> @btime (Profile.init(n=Int(1e8), delay=0.0000001); Profile.clear(); @profile myfunc())
  42.577 ms (2 allocations: 122.07 MiB)
0.9999998284031375
```
After (this PR):
```julia
julia> @btime (Profile.init(n=Int(1e8), delay=0.0001); Profile.clear(); @profile myfunc())
  41.659 ms (2 allocations: 122.07 MiB)
0.9999999728178411

julia> @btime (Profile.init(n=Int(1e8), delay=0.0000001); Profile.clear(); @profile myfunc())
  41.263 ms (2 allocations: 122.07 MiB)
0.9999999880968238
```

I also tried with `peakflops()` with varying delay sizes, and with different numbers of threads, and similarly never saw an impact. :)

---------

So, do i just merge this now? 😬 sorry still not sure about the etiquette now that i have commit rights! 😬